added Not Quite a Palindromic String

+++ Not Quite a Palindromic String/notQuiteaPalindromicString.cpp
@@ -2,11 +2,11 @@
 using namespace std;
 
 int main(){
-    int t;
+    long long t;
     cin >> t;
 
     while(t--){
-        int n;
+        long n;
         cin >> n;
 
         string s;
